requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
python-dotenv==1.0.0
mariadb==1.1.10
or-tools==9.9.3963
